requirements.txt
Django
djangorestframework
python-decouple
django-filter
django-autoslug
django-countries
Pillow
django-phonenumber-field
phonenumbers
psycopg2
flake8
black
isort
djoser
djangorestframework-simplejwt
PyJWT
pytest-django
pytest-factoryboy
Faker
pytest-cov
redis
celery
flower
django-celery-email
argh
PyYAML
watchdog
gunicorn==20.1.0